Complete rewrite of FootprintSpecTest.scala to conform to Specs conventions. Minor change of some test XML to correct a tag name: sponsoringOrganizationIDs --> sponsoringOrganizationsIDs
Minor bug fix in FootprintSpec.scala. The fromXML method in the VolunteerOpportunity companion object was incorrectly parsing
the contents of the sponsoringOrganizationsIDs tag, resulting the retrieved IDs being padded on both ends with spaces. This was
likely undetected as the mis-spelled tag in the testing code was keeping it from finding and sponsoringOrganizationsIDs.