Tickets #28, #30, #31 All are works-in-progress, but near completion.
src/main/webapp/static/front.html is a static page containing:
    - an inner template for a 50-50 column layout

They need to be separated and moved to their correct locations
(outside of the static directory).  I suppose once that's done,
there won't be much reason to keep a copy in the static
directory.  (Except to figure out why the page renders
differently in the static directory compared to when it comes
through the webapp setup.)